requirements.txt
gensim==4.0.1
tqdm==4.62.0
transformers==4.8.2
pytorch_ignite==0.4.4
imbalanced_learn==0.4.3
numpy==1.19.5
icecream==2.1.1
pandas==1.1.5
ignite==1.1.0
imblearn==0.0
scikit-learn=0.24.1